split out responsibilities
